ddl: support alter table .. truncate partition (#8624) (#9093)

+++ ddl/table.go
@@ -46,7 +46,7 @@ func onCreateTable(d *ddlCtx, t *meta.Meta, job *model.Job) (ver int64, _ error)
 	}
 
 	if tbInfo.Partition != nil {
-		err = checkAddPartitionTooManyPartitions(len(tbInfo.Partition.Definitions))
+		err = checkAddPartitionTooManyPartitions(uint64(len(tbInfo.Partition.Definitions)))
 		if err != nil {
 			job.State = model.JobStateCancelled
 			return ver, errors.Trace(err)
@@ -463,7 +463,7 @@ func onAddTablePartition(t *meta.Meta, job *model.Job) (ver int64, _ error) {
 		job.State = model.JobStateCancelled
 		return ver, errors.Trace(err)
 	}
-	err = checkAddPartitionTooManyPartitions(len(tblInfo.Partition.Definitions) + len(partInfo.Definitions))
+	err = checkAddPartitionTooManyPartitions(uint64(len(tblInfo.Partition.Definitions) + len(partInfo.Definitions)))
 	if err != nil {
 		job.State = model.JobStateCancelled
 		return ver, errors.Trace(err)
